bench/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "erreval-benchmark"
version = "1.0.0"
description = "ERR-EVAL: Epistemic Reasoning & Reliability Evaluation - An epistemic reliability benchmark"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Bennett Schwartz", email = "gc@gustycube.xyz" }
]
keywords = ["benchmark", "llm", "evaluation", "epistemic", "uncertainty"]

dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
erreval = "mirage.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["mirage"]
